Add option to delete temp file

+++ core/tasks.py
@@ -7,7 +7,21 @@ from celery import shared_task
 
 
 @shared_task
-def adoc_to_html(file_path, cache_key, content_type):
+def adoc_to_html(file_path, cache_key, content_type, delete_file=True):
+    """
+    Converts an AsciiDoc file to HTML and stores the result in the cache.
+    If delete_file is True, the temporary file will be deleted after the
+    conversion is complete.
+
+    The asciidoctor package is a Ruby gem, which is why we're using subprocess
+    to run the command.
+    https://docs.asciidoctor.org/asciidoctor/latest/
+
+    :param file_path: The path to the AsciiDoc file
+    :param cache_key: The key to use when storing the result in the cache
+    :param content_type: The content type of the file
+    :param delete_file: Whether or not to delete the temporary file after the
+    """
     result = subprocess.run(
         ["asciidoctor", "-o", "-", file_path],
         check=True,
@@ -27,6 +41,7 @@ def adoc_to_html(file_path, cache_key, content_type):
     )
 
     # Delete the temporary file
-    os.remove(file_path)
+    if delete_file:
+        os.remove(file_path)
 
     return converted_html
